caqtus_devices.orca_quest/src/caqtus_devices/orca_quest/configuration_editor/orca_quest_config_editor.py
from caqtus.gui.condetrol.device_configuration_editors._autogen import (
    build_device_configuration_editor,
)
from ..configuration import OrcaQuestCameraConfiguration

OrcaQuestConfigurationEditor = build_device_configuration_editor(
    OrcaQuestCameraConfiguration
)
