caqtus_devices.orca_quest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "caqtus-devices-orca-quest"
dynamic = ["version"]
description = "Contains the caqtus device extension for the Hamamatsu Orca Quest camera."
readme = "README.md"
requires-python = ">=3.12"
license = { file = "license.md" }
keywords = []
authors = [
    { name = "light-scattering-io", email = "lecaqtus66@gmail.com" },
]

dependencies = [
    "caqtus-suite>=6.14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/caqtus_devices"]

[tool.hatch.version]
path = "src/caqtus_devices/orca_quest/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "black>=24.3.0",
    "ruff>=0.6.4",
    "pre-commit>=3.7.0",
    "pyright>=1.1.384",
    "pytest>=8.1.1",
]
